Isolate deprecated `reference.genome` classes into a separate package
