+++ hpo-case-annotator-core/src/main/java/org/monarchinitiative/hpo_case_annotator/core/reference/genome/obsoleted/SingleFastaSequenceDao.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.hpo_case_annotator.core.reference.genome;
+package org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted;
 
 import htsjdk.samtools.SAMException;
 import htsjdk.samtools.reference.FastaSequenceIndex;
