+++ hpo-case-annotator-core/src/main/java/org/monarchinitiative/hpo_case_annotator/core/reference/genome/obsoleted/package-info.java
@@ -1,0 +1,2 @@
+@Deprecated
+package org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted;
