+++ hpo-case-annotator-core/src/test/java/org/monarchinitiative/hpo_case_annotator/core/reference/genome/obsoleted/SingleFastaSequenceDaoTest.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.hpo_case_annotator.core.reference.genome;
+package org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted;
 
 import htsjdk.samtools.SAMException;
 import org.junit.jupiter.api.AfterEach;
