+++ hpo-case-annotator-core/src/test/java/org/monarchinitiative/hpo_case_annotator/core/validation/ValidationRunnerTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.GenomeAssemblies;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.SequenceDao;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssemblies;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.SequenceDao;
 import org.monarchinitiative.hpo_case_annotator.model.proto.DiseaseCase;
 import org.monarchinitiative.hpo_case_annotator.model.proto.GenomeAssembly;
 import org.monarchinitiative.hpo_case_annotator.test.TestData;
