+++ hpo-case-annotator-core/src/test/resources/org/monarchinitiative/hpo_case_annotator/core/reference/genome/obsoleted/shortHg19ChromFa.tar.gz
