+++ hpo-case-annotator-forms/src/main/java/org/monarchinitiative/hpo_case_annotator/forms/v2/variant/unified/UnifiedVcfSymbolicVariantController.java
@@ -6,7 +6,7 @@ import javafx.beans.value.ChangeListener;
 import javafx.fxml.FXML;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.TextField;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.DeprecatedGenomicAssemblyRegistry;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.DeprecatedGenomicAssemblyRegistry;
 import org.monarchinitiative.hpo_case_annotator.forms.util.Formats;
 import org.monarchinitiative.hpo_case_annotator.forms.util.VariantTypeStringConverter;
 import org.monarchinitiative.svart.VariantType;
