+++ hpo-case-annotator-gui/src/main/java/org/monarchinitiative/hpo_case_annotator/gui/HpoCaseAnnotatorModule.java
@@ -6,6 +6,9 @@ import com.google.inject.Provides;
 import javafx.application.HostServices;
 import org.monarchinitiative.hpo_case_annotator.core.liftover.LiftOverAdapter;
 import org.monarchinitiative.hpo_case_annotator.core.reference.genome.*;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.DeprecatedGenomicAssemblyRegistry;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssemblies;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssembliesSerializer;
 import org.monarchinitiative.hpo_case_annotator.forms.*;
 import org.monarchinitiative.hpo_case_annotator.forms.GenomicAssemblyRegistry;
 import org.monarchinitiative.hpo_case_annotator.forms.v2.*;
