+++ hpo-case-annotator-gui/src/main/java/org/monarchinitiative/hpo_case_annotator/gui/controllers/SetResourcesController.java
@@ -10,7 +10,7 @@ import javafx.scene.control.TextField;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import org.monarchinitiative.hpo_case_annotator.core.io.EntrezParser;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.GenomeAssemblies;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssemblies;
 import org.monarchinitiative.hpo_case_annotator.gui.OptionalResources;
 import org.monarchinitiative.hpo_case_annotator.gui.io.Downloader;
 import org.monarchinitiative.hpo_case_annotator.gui.io.GenomeAssemblyDownloader;
