+++ hpo-case-annotator-gui/src/main/java/org/monarchinitiative/hpo_case_annotator/gui/util/StartupTask.java
@@ -3,7 +3,7 @@ package org.monarchinitiative.hpo_case_annotator.gui.util;
 import javafx.concurrent.Task;
 import org.monarchinitiative.hpo_case_annotator.core.io.EntrezParser;
 import org.monarchinitiative.hpo_case_annotator.core.io.OMIMParser;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.GenomeAssemblies;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssemblies;
 import org.monarchinitiative.hpo_case_annotator.gui.OptionalResources;
 import org.monarchinitiative.hpo_case_annotator.gui.Main;
 import org.slf4j.Logger;
