+++ hpo-case-annotator-gui/src/test/java/org/monarchinitiative/hpo_case_annotator/gui/TestHpoCaseAnnotatorModule.java
@@ -6,8 +6,8 @@ import javafx.application.HostServices;
 import org.monarchinitiative.hpo_case_annotator.core.io.EntrezParser;
 import org.monarchinitiative.hpo_case_annotator.core.io.OMIMParser;
 import org.monarchinitiative.hpo_case_annotator.core.liftover.LiftOverAdapter;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.GenomeAssemblies;
-import org.monarchinitiative.hpo_case_annotator.core.reference.genome.GenomeAssembliesSerializer;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssemblies;
+import org.monarchinitiative.hpo_case_annotator.core.reference.genome.obsoleted.GenomeAssembliesSerializer;
 import org.monarchinitiative.hpo_case_annotator.gui.controllers.*;
 import org.monarchinitiative.hpo_case_annotator.gui.controllers.variant.IntrachromosomalVariantController;
 import org.monarchinitiative.hpo_case_annotator.gui.controllers.variant.MendelianVariantController;
